requirements.txt
numpy==1.24.2
pandas==1.5.3
selenium==4.8.2
python-telegram-bot==20.1
